fix(apple): unwrap Node wrappers before writing env files
